requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-whois>=0.8.0
dnspython>=2.4.2
cryptography>=42.0.0
requests>=2.31.0
python-multipart>=0.0.6
geoip2>=4.8.0
aiohttp>=3.9.1
typing-extensions>=4.8.0
python-dotenv>=1.0.0
httpx>=0.25.2 